+++ app/routes.js
@@ -3787,6 +3787,9 @@ module.exports = function(app) {
 
 	// redirects everything else to our react app
 	app.get('*', function(req, res) {
+		res.setHeader('Cache-Control', 'no-cache, no-store, must-revalidate');
+		res.setHeader('Pragma', 'no-cache');
+		res.setHeader('Expires', '0');
 		res.sendFile(path.join(__dirname, '../ui-react/build/', 'index.html'));
 	});
 };